Fix: columnas de medida en staging a FLOAT (evita 'Numeric value out of range' de pyodbc fast_executemany en columnas DECIMAL)

+++ poc_colombia/schema/08_staging.sql
@@ -37,7 +37,7 @@ GO
 CREATE TABLE [stg].[tbl_Cuentas] (
     [Cod_Cuenta] INT NULL, [Cuenta] NVARCHAR(255) NULL, [Signo] INT NULL,
     [Agrupación] NVARCHAR(100) NULL, [Agrup_Gastos] INT NULL, [Cálculo] INT NULL,
-    [Por inmueble] INT NULL, [Divisor] DECIMAL(12,2) NULL, [Suma] INT NULL,
+    [Por inmueble] INT NULL, [Divisor] FLOAT NULL, [Suma] INT NULL,
     [Unidades] INT NULL, [Ranking_Inm] INT NULL, [Mostrar_pie] INT NULL,
     [Dispers_EjeX] INT NULL, [Dispers_EjeY] INT NULL, [Dispers_Burb] INT NULL
 );
@@ -70,7 +70,7 @@ GO
 CREATE TABLE [stg].[tbl_Contratos] (
     [FECHA] DATETIME2(0) NULL, [COD_FONDO] INT NULL, [ESTADO] NVARCHAR(30) NULL,
     [COD_CTR] NVARCHAR(20) NULL, [NIT] NVARCHAR(20) NULL, [NOM_ARREND] NVARCHAR(255) NULL,
-    [COD_INM] DECIMAL(18,0) NULL, [DET_INM] NVARCHAR(255) NULL, [GLA] DECIMAL(18,2) NULL,
+    [COD_INM] DECIMAL(18,0) NULL, [DET_INM] NVARCHAR(255) NULL, [GLA] FLOAT NULL,
     [Tipologia] NVARCHAR(100) NULL, [Fec_Inicio] DATETIME2(0) NULL, [Fec_Fin] DATETIME2(0) NULL,
     [IncremCanon] NVARCHAR(100) NULL
 );
@@ -79,62 +79,62 @@ GO
 /* -- Facts "Valores" (mismo layout, ver 04_facts_valores.sql) -- */
 CREATE TABLE [stg].[tbl_Valores] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Fondo] INT NULL,
-    [Cod_Nivel] INT NULL, [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Nivel] INT NULL, [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_Valor_Libros] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_Gastos_Otros] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Fondo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_Gastos] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_Avaluos] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_Ingresos] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_Nexus] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Fondo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_Viva_Malls] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_GLA_2008_2023] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_2011_2023_PYG] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Valores_2011_2023_VlrLibros] (
     [Fecha] DATETIME2(0) NULL, [Cod_Tiempo] INT NULL, [Cod_Nivel] INT NULL,
-    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Centro] NVARCHAR(20) NULL, [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Pagos_Inmuebles] (
     [Fecha] DATETIME2(0) NULL, [Fecha_TRN] DATETIME2(0) NULL, [Cod_Inm] DECIMAL(18,0) NULL,
-    [Cod_Cuenta] INT NULL, [Valor] DECIMAL(18,4) NULL
+    [Cod_Cuenta] INT NULL, [Valor] FLOAT NULL
 );
 GO
 
@@ -155,36 +155,36 @@ CREATE TABLE [stg].[tbl_Cuentas_EEFF] (
 GO
 CREATE TABLE [stg].[tbl_EEFF] (
     [FECHA] DATETIME2(0) NULL, [COD_FONDO] INT NULL, [COD_CTA] DECIMAL(18,0) NULL,
-    [NIT_TERCERO] DECIMAL(18,0) NULL, [SALDO] DECIMAL(18,4) NULL
+    [NIT_TERCERO] DECIMAL(18,0) NULL, [SALDO] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[F351] (
     [FECHA] DATETIME2(0) NULL, [INMUEBLE] NVARCHAR(255) NULL, [matrícula] NVARCHAR(50) NULL,
     [Unidad de Captura] DECIMAL(18,0) NULL, [No asignado por la entidad] DECIMAL(18,0) NULL,
-    [Fecha emisión] FLOAT NULL, [Valor nominal] DECIMAL(19,4) NULL,
-    [Valor de compra moneda original] DECIMAL(19,4) NULL, [Valor de compra en pesos] DECIMAL(19,4) NULL,
-    [Vr mercado o valor presente en $] DECIMAL(19,4) NULL, [Campo11] NVARCHAR(255) NULL,
-    [Campo12] FLOAT NULL, [Campo13] DECIMAL(19,4) NULL, [Campo14] DECIMAL(19,4) NULL, [F15] FLOAT NULL
+    [Fecha emisión] FLOAT NULL, [Valor nominal] FLOAT NULL,
+    [Valor de compra moneda original] FLOAT NULL, [Valor de compra en pesos] FLOAT NULL,
+    [Vr mercado o valor presente en $] FLOAT NULL, [Campo11] NVARCHAR(255) NULL,
+    [Campo12] FLOAT NULL, [Campo13] FLOAT NULL, [Campo14] FLOAT NULL, [F15] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[VL_CentralPoint] (
-    [FECHA] DATETIME2(0) NULL, [VLR_FINAL_E1] DECIMAL(18,4) NULL, [VLR_FINAL_E2] DECIMAL(18,4) NULL,
-    [VLR_E1] DECIMAL(18,4) NULL, [VLR_E2] DECIMAL(18,4) NULL, [MVA] DECIMAL(18,4) NULL,
-    [TOTAL_ACTIVOS_PA] DECIMAL(18,4) NULL, [PASIVO_PA] DECIMAL(18,4) NULL, [ANT_DF] DECIMAL(18,4) NULL,
-    [TOTAL] DECIMAL(18,4) NULL, [351] DECIMAL(18,4) NULL, [DIF] DECIMAL(18,4) NULL,
-    [UVR_E1] DECIMAL(18,4) NULL, [UVR_E2] DECIMAL(18,4) NULL, [UVR_E1_#2] DECIMAL(18,4) NULL,
-    [UVR_E2_#2] DECIMAL(18,4) NULL, [AVAL_E1] DECIMAL(18,4) NULL, [AVAL_E2] DECIMAL(18,4) NULL,
-    [VLR_AVAL_E1] DECIMAL(18,4) NULL, [VLR_AVAL_E2] DECIMAL(18,4) NULL
+    [FECHA] DATETIME2(0) NULL, [VLR_FINAL_E1] FLOAT NULL, [VLR_FINAL_E2] FLOAT NULL,
+    [VLR_E1] FLOAT NULL, [VLR_E2] FLOAT NULL, [MVA] FLOAT NULL,
+    [TOTAL_ACTIVOS_PA] FLOAT NULL, [PASIVO_PA] FLOAT NULL, [ANT_DF] FLOAT NULL,
+    [TOTAL] FLOAT NULL, [351] FLOAT NULL, [DIF] FLOAT NULL,
+    [UVR_E1] FLOAT NULL, [UVR_E2] FLOAT NULL, [UVR_E1_#2] FLOAT NULL,
+    [UVR_E2_#2] FLOAT NULL, [AVAL_E1] FLOAT NULL, [AVAL_E2] FLOAT NULL,
+    [VLR_AVAL_E1] FLOAT NULL, [VLR_AVAL_E2] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[VL_Disp_xa_Venta] (
     [FECHA] DATETIME2(0) NULL, [COD_INM] DECIMAL(18,0) NULL, [DESCR_INM] NVARCHAR(255) NULL,
-    [VALOR] DECIMAL(18,4) NULL
+    [VALOR] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_ValorLibros_xInmueble] (
     [DESCR] NVARCHAR(255) NULL, [FECHA] DATETIME2(0) NULL, [Cod_Inm] DECIMAL(18,0) NULL,
-    [COD_CTA] DECIMAL(18,0) NULL, [VALOR] DECIMAL(18,4) NULL
+    [COD_CTA] DECIMAL(18,0) NULL, [VALOR] FLOAT NULL
 );
 GO
 CREATE TABLE [stg].[tbl_Cruce_SIF] (
